% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_amazon.R
\docType{data}
\name{amazon}
\alias{amazon}
\title{Amazon experiment dataset}
\format{
A data frame with 3000 rows and 6 variables:
\describe{
  \item{logo}{Factor with three values: 1 
  (= the original), 2, 3}
  \item{after_sales}{annual purchase of a particular
   customer in USD after the experiment is implemented}
  \item{before_sales}{annual purchase of a particular
   customer in USD before the experiment is implemented}
  \item{country}{0 = US; 1 = UK}
  \item{age}{age of customer}
  \item{sex}{Factor, with Male = 0 and Female = 1}
}
}
\source{
Plato Leung \email{y.k.leung@jads.nl}
}
\description{
A dataset collected from an Amazon experiment.
}
\details{
The experiment relates to the change of a logo 
by Amazon. In fact, three logos were used (logo1, 
logo2, logo3) and the interest is to test how  
the amount consumers spend is influenced by each 
logo and which logo has the most positive effect 
(if any). 

This is a typical example of so-called *A-B testing*, 
where marketeers try out several conditions (here: 
logos) in order to check which condition works best.

Ask Plato Leung for more info about the dataset.
}
\keyword{datasets}
