% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterpairs.r
\name{betterpairs}
\alias{betterpairs}
\title{Pimped scatterplot matrix}
\usage{
betterpairs(
  x,
  histogram = FALSE,
  density = TRUE,
  rug = TRUE,
  smoother = TRUE,
  hist.col = "red",
  method = "pearson",
  cex.cor = NULL,
  ...
)
}
\arguments{
\item{x}{a matrix or dataframe}

\item{histogram}{logical, Should a histogram be plotted on the diagonal (or only variable names)}

\item{density}{logical, Should a histogram+density plot be plotted on the diagonal (or only variable names). 
If \code{histogram=TRUE} the value for \code{density} is not used, a histogram is plotted}

\item{rug}{logical, should \code{rug} be included in the histogram? Default = TRUE.}

\item{smoother}{logical, should 'smoother' be used below the diagonal (default is TRUE). Otherwise \code{panel.smooth} is used.}

\item{hist.col}{color of the histogram}

\item{method}{method for calculating the correlation (default = \code{pearson})}

\item{cex.cor}{(optioneel) numeriek, schaalt de grootte van de text waarmee de correlatie wordt weergegeven. 
Heeft geen effect op de grootte van de asterisks. Indien \code{NULL} (de default) kan het verschil in grootte erg groot worden.}

\item{...}{additional arguments, to be passed to \code{\link[graphics]{pairs}}}
}
\value{
Scatterplot matrix with histograms on the diagonal, 
scatterplots with smoothers below the diagonal and
pairwise correlations (with signif. stars) above the diagonal
}
\description{
Makes a pimped scatterplot matrix
}
\details{
This is a pimped version of the standard \code{\link[graphics]{pairs}} plot. 
If you want the original, more boring version, go there. 
This function is a lot cooler, though.

For datasets with lots of variables, it is usually best to look at 
the pairs plots for subsets of variables, unless you have a really large 
monitor...
}
\examples{
t <- matrix(runif(200), 40, 5)
colnames(t) <- LETTERS[1:5]
betterpairs(t)

\dontrun{
data(movie, package = "bootcamp2021")
}
}
