% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootcamp2021-package.r
\docType{package}
\name{bootcamp2021}
\alias{bootcamp2021}
\title{bootcamp2021}
\description{
Package supporting the bootcamp2021 course
}
\details{
Functions, datasets, and tutorials, useful for the bootcamp2021 course.
}
\keyword{package}
