% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootcamp_2021-package.r
\docType{package}
\name{bootcamp_2021}
\alias{bootcamp_2021}
\title{bootcamp_2021}
\description{
Package supporting the bootcamp_2021 course
}
\details{
Functions, datasets, and tutorials, useful for the bootcamp_2021 course.
}
\keyword{package}
