% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brendonSmall.R
\docType{data}
\name{BrendonSmall}
\alias{BrendonSmall}
\title{Hypothetical data for Brendon Small and company}
\source{
\url{http://rcompanion.org/handbook/I_10.html}
}
\description{
A data frame of the intake of calories and sodium for
 students in five classes. Hypothetical data.
}
\keyword{datasets}
