% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_checks.R
\name{check_packages}
\alias{check_packages}
\title{Check for presence of packages}
\usage{
check_packages(reqs)
}
\arguments{
\item{reqs}{data.frame (see details)}
}
\value{
character string, summarizing the results
}
\description{
Checks whether a set of packages is present and of the correct version
}
\details{
Useful inside a learnr tutorial.

\code{reqs} is a data.frame with column names "pkg", "version", "where". 
All items are \code{character}! 

\itemize{
\item{pkg}{names of the packages}
\item{version}{the minimally required version}
\item{where}{location of the package for download}
}

For \code{where} there are two options. If the package resides on CRAN, 
it should be "CRAN". If it is on github, it should be "username/reponame" (ie. 
whatever would go inside \code{remotes::install_github(where)}).

First, the function checks if a package is installed and, if installed, whether 
it has the required version number or higher. 
If eigher of these checks fail, it will attempt to download/upgrade the package.

After a first pass along all packages (as described above), a second check is 
performed if a package is now installed and, if installed, whether 
it has the required version number or higher. 

If a package is still not installed, a message is returned that tells the user 
how to install it manually. 
If a package still does not at least have the required version, 
a message is returned that tells the user how to upgrade it manually.
}
\examples{
\dontrun{
pkgs <- matrix(c(
"gradethis", "0.2.3.9001", "rstudio/gradethis",
"igraph", "1.2.6", "CRAN",
"influenceR", "0.1.0", "CRAN",
"intergraph", "2.0-2", "CRAN",
"network", "1.17.1", "CRAN",
"networkD3", "0.4", "CRAN",
"sna", "2.6", "CRAN",
"SNA4DSData", "0.9.9000", "SNAnalyst/SNA4DSData"
), byrow = TRUE, ncol = 3) |> 
  as.data.frame() |> 
  setNames(c("pkg", "version", "where"))
  
check_packages(pkgs)
}
}
\keyword{internal}
