% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_checks.R
\name{check_r_equal}
\alias{check_r_equal}
\title{Check for correct version of R}
\usage{
check_r_equal(Major = 4, Minor = 1.1)
}
\arguments{
\item{Major}{Major version number, e.g., 4}

\item{Minor}{Minor version number}
}
\value{
message with the result
}
\description{
Check for correct version of R
}
\details{
Checks whether the user is using the correct version of R (exactly). 
Note that this checks for the exact version: if a user has a higher version of 
R, the function will report that the user is not using the correct version.
}
\keyword{internal}
