% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_helpers.R
\name{compute_biserial_correlation}
\alias{compute_biserial_correlation}
\title{Bereken biseriale correlatie}
\usage{
compute_biserial_correlation(x, y)
}
\arguments{
\item{x}{numerieke vector (continu)}

\item{y}{numerieke vector (binair, gecodeerd als 0/1)}
}
\value{
lijst met $estimate, $type, $n, $p.value, $statistic, $method
}
\description{
Berekent de biseriale correlatie tussen een continue variabele \code{x}
en een binaire variabele \code{y}.
}
\details{
De biseriale correlatie corrigeert de punt-biseriale correlatie (= Pearson)
voor de kunstmatige dichotomisering van een onderliggende continue,
normaal verdeelde variabele. De formule is:

\deqn{r_{bis} = r_{pb} \cdot \frac{\sqrt{p \cdot q}}{\phi(z_p)}}

waar:
\itemize{
  \item \eqn{r_{pb}} = punt-biseriale correlatie (= Pearson correlatie)
  \item \eqn{p} = proportie observaties met y=1
  \item \eqn{q = 1 - p}
  \item \eqn{\phi(z_p)} = de dichtheid van de standaard normaal op het punt
    \eqn{z_p = \Phi^{-1}(p)}
}

AANNAME: de onderliggende continue variabele die gedichtomiseerd is tot y
volgt een normale verdeling. Als deze aanname niet klopt, is de biseriale
correlatie niet zinvol.

OPMERKING: de biseriale correlatie kan in theorie buiten [-1, 1] vallen
door de correctiefactor. We clampen het resultaat naar [-1, 1].
}
\keyword{internal}
