% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Univariate descriptives}
\usage{
descriptives(
  x,
  na.rm = TRUE,
  normal = FALSE,
  ranges = TRUE,
  trim = 0.1,
  fast = NULL,
  quantiles = NULL,
  IQR = FALSE,
  omit_nonnumeric = TRUE,
  complete_cases = FALSE,
  digits = 2,
  print = TRUE
)
}
\arguments{
\item{x}{A data frame, vector, or matrix}

\item{na.rm}{logical, \code{na.rm = FALSE} will delete the case in the 
computation of statistics such as \code{mean}, \code{sd}, et cetera}

\item{normal}{logical, if \code{TRUE} skewness and kurtosis are computed and the Shapiro-Wilk test of normality.}

\item{ranges}{logical, if \code{TRUE} (the default) min, max, range, 
mad, median are computed}

\item{trim}{numeric between 0 and 1 (default trim = .1) –- drops the top and 
bottom trim fraction in the computation of the mean}

\item{fast}{if TRUE, will do fewer statistics for an improvement in speed. 
If default \code{NULL}, will switch to fast mode for large 
(ncol * nrow > 10^7) 
problems, otherwise defaults to \code{fast = FALSE}}

\item{quantiles}{if not \code{NULL}, will find the specified quantiles 
(e.g. \code{quant = c(.25, .75)} will find the 25th and 75th percentiles)}

\item{IQR}{logical, if \code{TRUE} (default is \code{FALSE}) the interquartile 
range is computed}

\item{omit_nonnumeric}{logical, should non-numeric variables be dropped? 
Defaults to \code{TRUE} because most desciptives in this function are not 
meaningful for factors or characters.}

\item{complete_cases}{logical, should only complete cases be used? 
If \code{TRUE}, all rows in \code{x} that have a \code{NA} in them will be 
completely dropped. Only useful if this is the intended behavior in the 
analysis that will be run with the data. Default is \code{FALSE}.}

\item{digits}{numeric, the number of decimals to print}

\item{print}{logical, should the descriptives be printed to the console? 
Defaults to \code{TRUE}.}
}
\value{
optionally prints the result to the console. The resulting data.frame 
is invisibly returned.
}
\description{
Determine univariate descriptives
}
\details{
Determines a series of univariate statistics for the variables in a dataset. 
The result is printed if \code{print} is \code{TRUE}. It is also invisibly 
returned, so it is possible to not print the result, but only use the 
descriptives data.frame object for further handling.

Currently implemented:

\describe{
\item{column}{column number of the variable inside \code{x}}
\item{n_valid}{number of non-missing observations}
\item{n_na}{number of \code{NA}'s}
\item{mean}{mean}
\item{trimmed_mean}{trimmed mean}
\item{sd}{standard deviation}
\item{se}{standard error}
\item{coefvar}{ceofficient of variation: sd/mean}
\item{skewness}{skewness}
\item{kurtosis}{kurtosis}
\item{normal.w}{Shapiro-Wilk test of normality}
\item{normal.p}{p-value of Shapiro-Wilk test of normality}
\item{min}{minimum}
\item{max}{maximum}
\item{median}{median}
\item{range}{max - min}
\item{mad}{mean absolute deviation}
\item{quantiles}{the specified quantile levels}
\item{IQR}{interquartile range Q0.75 - Q0.25}
}
}
\examples{
descriptives(bootcamp2021::movie)
descriptives(bootcamp2021::movie, omit_nonnumeric = FALSE)
descriptives(bootcamp2021::movie, digits = 3)

result <- descriptives(bootcamp2021::movie, print = FALSE)
print(result)

descriptives(bootcamp2021::movie, quantiles = c(.1, .9), IQR = TRUE)
}
