% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dozen.R
\docType{data}
\name{dozen}
\alias{dozen}
\title{Twelve datasets}
\description{
A list with twelve datasets
}
\details{
\code{dozen} is a list with twelve datasets, each 
consisting of two variables: \code{x} and {y}.

The datasets are named "one", "two", ..., ""twelve".

Useful to show the value of visualization, next 
to numeric summaries.

Each dataset has 142 rows and 2 columns.

Also known as the `datasauRus dozen long`.
}
\keyword{datasets}
