% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis computation}
\usage{
kurtosis(x, na.rm = TRUE)
}
\arguments{
\item{x}{a \code{variable}, \code{matrix} or a \code{data.frame}}

\item{na.rm}{logical, should \code{NA}s be removed before
computation?}
}
\value{
numeric
}
\description{
Calculates the kurtosis of a variable
}
\details{
Calculates kurtosis coefficient for given variable, 
\code{matrix}, or \code{data.frame}.

Kurtosis is a measure of whether the data are heavy-tailed or
light-tailed relative to a normal distribution.
That is, data sets with high kurtosis tend to have heavy tails, or
outliers. Data sets with low kurtosis tend to have light tails, or
lack of outliers. A uniform distribution would be the extreme case.

The maximum kurtosis leads to this measure being close to 
n (=the number of observations). Minimal kurtosis is 0.
}
\examples{
kurtosis(rnorm(100)) # about 0, minimal kurtosis

# slightly heavier tail
kurtosis(rt(100, 10))# approx .35
# quite heavy tail
kurtosis(rt(100, 1))# approx 84

# maximal kurtosis
kurtosis(c(rep(10, times = 1000), 101)) # 994.006
}
