% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MediaElectionData.R
\docType{data}
\name{MediaElectionData}
\alias{MediaElectionData}
\title{Media Election Data}
\format{
A data frame with 2509 rows and 9 variables:
\describe{
  \item{TVtime}{}
  \item{ONLINEtime}{}
  \item{PolInt}{}
  \item{ElecInt}{}
  \item{ProbLeft}{}
  \item{ProbRight}{}
  \item{age}{age of respondents}
  \item{gender}{Factor, with Male = 0 and Female = 1}
  \item{edu}{}
}
}
\source{
Claudia Zucca \email{c.zucca@jads.nl}
}
\description{
A dataset derived from the British Election Study (BES)
}
\details{
A few variables of relevance have been selected from the BEF
and recoded for teaching purposes
}
\keyword{datasets}
