% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_corr.R
\name{partial_corr}
\alias{partial_corr}
\title{Partial correlation matrix}
\usage{
partial_corr(
  x,
  use = "everything",
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{correlation matrix or dataset in the form of a \code{matrix} or \code{data.frame}}

\item{use}{an optional character string giving a method for computing 
covariances in the presence of missing values. This must be (an abbreviation of) 
one of the strings \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"}, 
\code{"na.or.complete"}, or \code{"pairwise.complete.obs"}}

\item{method}{a character string indicating which correlation coefficient 
(or covariance) is to be computed. One of \code{"pearson"} (default), 
\code{"kendall"}, or \code{"spearman"}: can be abbreviated}
}
\value{
the matrix or partial correlations
}
\description{
Determine a partial correlation matrix
}
\details{
\code{partial_corr} computes the pairwise *partial correlation* coefficients 
from either a correlation matrix or a data matrix.

The approach is to find the complete partial correlation matrix 
(that is, partial all the other variables out of each variable). 
This may be done by simply specifying the raw data or correlation matrix. 
(In the case of raw data, correlations will be found according to \code{use} 
and \code{method}.) In this case, just specify the data matrix.
This is useful in the case of multiple regression.
}
\examples{
partial_corr(bootcamp2021::movie[, -3])
partial_corr(bootcamp2021::centrality)
}
