% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_helpers.R
\name{pbivnorm}
\alias{pbivnorm}
\title{Bivariate normale cumulatieve verdelingsfunctie}
\usage{
pbivnorm(x, y, rho)
}
\arguments{
\item{x}{numeriek, bovengrens voor de eerste variabele}

\item{y}{numeriek, bovengrens voor de tweede variabele}

\item{rho}{numeriek in [-1, 1], de correlatie}
}
\value{
numeriek, de bivariate normale CDF-waarde P(X <= x, Y <= y)
}
\description{
Berekent P(X <= x, Y <= y) voor een standaard bivariate normale verdeling
met correlatie rho, via het Drezner-Wesolowsky (1990) algoritme.
}
\details{
Het algoritme gebruikt 20-punts Gauss-Legendre kwadratuur van de
integraalkernel.

Referentie: Drezner, Z. & Wesolowsky, G.O. (1990). On the computation of
the bivariate normal integral. Journal of Statistical Computation and
Simulation, 35, 101-107.
}
\keyword{internal}
