% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_helpers.R
\name{pbivnorm_low_corr}
\alias{pbivnorm_low_corr}
\title{Bivariate normaal CDF voor lage correlatie (|rho| < 0.925)}
\usage{
pbivnorm_low_corr(x, y, rho, w, t_nodes)
}
\arguments{
\item{x, y}{bovengrenzen}

\item{rho}{correlatie}

\item{w}{kwadratuurgewichten}

\item{t_nodes}{kwadratuurknopen op [0, 1]}
}
\value{
numerieke CDF-waarde
}
\description{
Directe Gauss-Legendre kwadratuur van de integraalkernel.
De transformatie brengt de integratie over [0, rho] naar een
gewogen som van functie-evaluaties.
}
\keyword{internal}
