% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, na.rm = TRUE)
}
\arguments{
\item{x}{a \code{variable}, \code{matrix} or a \code{data.frame}}

\item{na.rm}{logical, should \code{NA}s be removed before
computation?}
}
\value{
numeric
}
\description{
Calculates the skewness of a variable
}
\details{
Skewness is a measure of symmetry, or more precisely, the lack of symmetry. A distribution, or data set, is symmetric if it looks the same to the left and right of the center point.

The skewness for a normal distribution is zero, and any symmetric
data should have a skewness near zero. Negative values for the
skewness indicate data that are skewed left and positive values for
the skewness indicate data that are skewed right. By skewed left, we
mean that the left tail is long relative to the right tail. Similarly,
skewed right means that the right tail is long relative to the left
tail. If the data are multimodal, then this may affect the sign of
the skewness.

For left skewed data, you get a negative value for \code{skewness}. 
For right skewed data, you get a positive value for \code{skewness}.
}
\examples{
data <- c(rep(10, 3), rep(15, 7), rep(25, 9), 
rep(35, 6), rep(45, 6), rep(55, 2), rep(65, 1), rep(95, 1))

# right-skewed data
hist(data)
skewness(data) # 1.277999
kurtosis(data) # 2.169775

# left-skewed data
data0 <- -data
hist(data0)
skewness(data0) # -1.277999
kurtosis(data0) # 2.169775

skewness(rnorm(100)) # approx 0
}
