% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_states.R
\docType{data}
\name{states}
\alias{states}
\title{states dataset}
\format{
A data frame with 50 rows and 10 variables:
\describe{
  \item{Population}{population estimate as of July 1, 1975}
  \item{Income}{per capita income (1974)}
  \item{Illiteracy}{illiteracy (1970, percent of population)}
  \item{Life Exp}{life expectancy in years (1969–71)}
  \item{Murder}{murder and non-negligent manslaughter rate 
  per 100,000 population (1976)}
  \item{HS Grad}{percent high-school graduates (1970)}
  \item{Frost}{mean number of days with minimum temperature below freezing
   (1931–1960) in capital or large city}
 \item{Area}{land area in square miles}
 \item{Name}{name of the state}
 \item{Region}{factor giving the region (Northeast, South, North Central, West) that each state belongs to}
}
}
\source{
\code{states <- as.data.frame(state.x77)}
\code{states$Name <- rownames(states)}
\code{states$region <- state.region}
\code{sapply(states, class)}
}
\description{
A dataset related to the 50 states of the United States of America.
}
\details{
This is a dataset that joins the \code{state.x77} and 
\code{state.region}.
}
\keyword{datasets}
