% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_checks.R
\name{version_check}
\alias{version_check}
\alias{check_rstudio_equal}
\alias{check_r_equal}
\alias{check_rstudio_equal_or_larger}
\alias{check_r_equal_or_larger}
\title{Check for correct version of R or RStudio}
\usage{
check_rstudio_equal(version = "2023.6.1", verdict = TRUE)

check_r_equal(version = "4.3.1", verdict = TRUE)

check_rstudio_equal_or_larger(version = "2023.6.1.524", verdict = TRUE)

check_r_equal_or_larger(version = "4.3.1", verdict = TRUE)
}
\arguments{
\item{version}{required version number to test against}

\item{verdict}{logical, whether a text with the verdict needs to be printed
to the console}
}
\value{
logical (invisibly) and, if asked for, a printed verdict
}
\description{
Check for correct version of R or RStudio
}
\details{
Checks whether the user is using the correct version of RStudio or R

The result is a logical (\code{TRUE} or \code{FALSE}), which is returned
invisibly.

If required, a verdict can be printed to the console as well (when \code{verdict = TRUE})

The functions \code{check_rstudio_equal} and  \code{check_r_equal} check
whether RStudio or R have the exact version asked for.

The functions \code{check_rstudio_equal_or_larger} and \code{check_r_equal_or_larger} check
whether RStudio or R have at least the version asked for (havig a higher version is fine).
}
\section{Functions}{
\itemize{
\item \code{check_rstudio_equal()}: Check whether RStudio has exactly the requested version.

\item \code{check_r_equal()}: Check whether R has exactly the requested version.

\item \code{check_rstudio_equal_or_larger()}: Check whether RStudio has at least the requested version.

\item \code{check_r_equal_or_larger()}: Check whether R has at least the requested version.

}}
\keyword{internal}
